feat: add SCL automatic approach for Navigator1, refactor: sample app by moving widgets in main.dart to separate files
