+++ example/lib/src/components/page.dart
@@ -1,0 +1,35 @@
+part of '../../main.dart';
+
+class Page extends StatelessWidget {
+  final String title;
+  final GlobalKey<ScaffoldState>? scaffoldKey;
+  final List<Widget> children;
+  final Widget? floatingActionButton;
+  final FloatingActionButtonLocation? floatingActionButtonLocation;
+
+  const Page({
+    Key? key,
+    required this.title,
+    this.scaffoldKey,
+    this.floatingActionButton,
+    this.floatingActionButtonLocation,
+    required this.children,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      key: scaffoldKey,
+      appBar: AppBar(title: Text(title)),
+      body: SingleChildScrollView(
+          physics: const ClampingScrollPhysics(),
+          padding: const EdgeInsets.only(top: 20.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: children,
+          )),
+      floatingActionButton: floatingActionButton,
+      floatingActionButtonLocation: floatingActionButtonLocation,
+    );
+  }
+}
