+++ example/lib/src/screens/my_home_page.dart
@@ -1,0 +1,332 @@
+part of '../../main.dart';
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  final buttonStyle = ButtonStyle(
+    backgroundColor: MaterialStateProperty.all(Colors.lightBlue),
+    foregroundColor: MaterialStateProperty.all(Colors.white),
+  );
+
+  List<ReportType> reportTypes = [];
+
+  final primaryColorController = TextEditingController();
+  final screenNameController = TextEditingController();
+
+  void restartInstabug() {
+    Instabug.setEnabled(false);
+    Instabug.setEnabled(true);
+    BugReporting.setInvocationEvents([InvocationEvent.floatingButton]);
+  }
+
+  void setOnDismissCallback() {
+    BugReporting.setOnDismissCallback((dismissType, reportType) {
+      showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            title: const Text('On Dismiss'),
+            content: Text(
+              'onDismiss callback called with $dismissType and $reportType',
+            ),
+          );
+        },
+      );
+    });
+  }
+
+  void show() {
+    Instabug.show();
+  }
+
+  void reportScreenChange() {
+    Instabug.reportScreenChange(screenNameController.text);
+  }
+
+  void sendBugReport() {
+    BugReporting.show(ReportType.bug, [InvocationOption.emailFieldOptional]);
+  }
+
+  void sendFeedback() {
+    BugReporting.show(
+        ReportType.feedback, [InvocationOption.emailFieldOptional]);
+  }
+
+  void showNpsSurvey() {
+    Surveys.showSurvey('pcV_mE2ttqHxT1iqvBxL0w');
+  }
+
+  void showManualSurvey() {
+    Surveys.showSurvey('PMqUZXqarkOR2yGKiENB4w');
+  }
+
+  final _scaffoldKey = GlobalKey<ScaffoldState>();
+
+  void getCurrentSessionReplaylink() async {
+    final result = await SessionReplay.getSessionReplayLink();
+    if (result == null) {
+      const snackBar = SnackBar(
+        content: Text('No Link Found'),
+      );
+      ScaffoldMessenger.of(_scaffoldKey.currentContext!).showSnackBar(snackBar);
+    } else {
+      var snackBar = SnackBar(
+        content: Text(result),
+      );
+      ScaffoldMessenger.of(_scaffoldKey.currentContext!).showSnackBar(snackBar);
+    }
+  }
+
+  void showFeatureRequests() {
+    FeatureRequests.show();
+  }
+
+  void toggleReportType(ReportType reportType) {
+    if (reportTypes.contains(reportType)) {
+      reportTypes.remove(reportType);
+    } else {
+      reportTypes.add(reportType);
+    }
+    BugReporting.setReportTypes(reportTypes);
+  }
+
+  void changeFloatingButtonEdge() {
+    BugReporting.setFloatingButtonEdge(FloatingButtonEdge.left, 200);
+  }
+
+  void setInvocationEvent(InvocationEvent invocationEvent) {
+    BugReporting.setInvocationEvents([invocationEvent]);
+  }
+
+  void changePrimaryColor() {
+    String text = 'FF' + primaryColorController.text.replaceAll('#', '');
+    Color color = Color(int.parse(text, radix: 16));
+    Instabug.setPrimaryColor(color);
+  }
+
+  void setColorTheme(ColorTheme colorTheme) {
+    Instabug.setColorTheme(colorTheme);
+  }
+
+  void _navigateToCrashes() {
+    ///This way of navigation utilize screenLoading automatic approach [Navigator 1]
+    Navigator.pushNamed(context, CrashesPage.screenName);
+
+    ///This way of navigation utilize screenLoading manual approach [Navigator 1]
+    // Navigator.push(
+    //   context,
+    //   MaterialPageRoute(
+    //     builder: (context) => const CrashesPage(),
+    //     settings: const RouteSettings(name: CrashesPage.screenName),
+    //   ),
+    // );
+  }
+
+
+  void _navigateToApm() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const InstabugCaptureScreenLoading(
+          screenName: ApmPage.screenName,
+          child: ApmPage(),
+        ),
+        settings: const RouteSettings(name: ApmPage.screenName),
+      ),
+    );
+  }
+
+  void _navigateToComplex() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const ComplexPage(),
+        settings: const RouteSettings(name: ComplexPage.screenName),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Page(
+      scaffoldKey: _scaffoldKey,
+      title: widget.title,
+      children: [
+        Container(
+          margin: const EdgeInsets.only(left: 20.0, right: 20.0, bottom: 20.0),
+          child: const Text(
+            'Hello Instabug\'s awesome user! The purpose of this application is to show you the different options for customizing the SDK and how easy it is to integrate it to your existing app',
+            textAlign: TextAlign.center,
+          ),
+        ),
+        InstabugButton(
+          onPressed: restartInstabug,
+          text: 'Restart Instabug',
+        ),
+        const SectionTitle('Primary Color'),
+        InstabugTextField(
+          controller: primaryColorController,
+          label: 'Enter primary color',
+        ),
+        InstabugButton(
+          text: 'Change Primary Color',
+          onPressed: changePrimaryColor,
+        ),
+        const SectionTitle('Change Invocation Event'),
+        ButtonBar(
+          mainAxisSize: MainAxisSize.min,
+          alignment: MainAxisAlignment.start,
+          children: <Widget>[
+            ElevatedButton(
+              onPressed: () => setInvocationEvent(InvocationEvent.none),
+              style: buttonStyle,
+              child: const Text('None'),
+            ),
+            ElevatedButton(
+              onPressed: () => setInvocationEvent(InvocationEvent.shake),
+              style: buttonStyle,
+              child: const Text('Shake'),
+            ),
+            ElevatedButton(
+              onPressed: () => setInvocationEvent(InvocationEvent.screenshot),
+              style: buttonStyle,
+              child: const Text('Screenshot'),
+            ),
+          ],
+        ),
+        ButtonBar(
+          mainAxisSize: MainAxisSize.min,
+          alignment: MainAxisAlignment.start,
+          children: <Widget>[
+            ElevatedButton(
+              onPressed: () =>
+                  setInvocationEvent(InvocationEvent.floatingButton),
+              style: buttonStyle,
+              child: const Text('Floating Button'),
+            ),
+            ElevatedButton(
+              onPressed: () =>
+                  setInvocationEvent(InvocationEvent.twoFingersSwipeLeft),
+              style: buttonStyle,
+              child: const Text('Two Fingers Swipe Left'),
+            ),
+          ],
+        ),
+        InstabugButton(
+          onPressed: show,
+          text: 'Invoke',
+        ),
+        InstabugButton(
+          onPressed: setOnDismissCallback,
+          text: 'Set On Dismiss Callback',
+        ),
+        const SectionTitle('Repro Steps'),
+        InstabugTextField(
+          controller: screenNameController,
+          label: 'Enter screen name',
+        ),
+        InstabugButton(
+          text: 'Report Screen Change',
+          onPressed: reportScreenChange,
+        ),
+        InstabugButton(
+          onPressed: sendBugReport,
+          text: 'Send Bug Report',
+        ),
+        InstabugButton(
+          onPressed: showManualSurvey,
+          text: 'Show Manual Survey',
+        ),
+        const SectionTitle('Change Report Types'),
+        ButtonBar(
+          mainAxisSize: MainAxisSize.min,
+          alignment: MainAxisAlignment.start,
+          children: <Widget>[
+            ElevatedButton(
+              onPressed: () => toggleReportType(ReportType.bug),
+              style: buttonStyle,
+              child: const Text('Bug'),
+            ),
+            ElevatedButton(
+              onPressed: () => toggleReportType(ReportType.feedback),
+              style: buttonStyle,
+              child: const Text('Feedback'),
+            ),
+            ElevatedButton(
+              onPressed: () => toggleReportType(ReportType.question),
+              style: buttonStyle,
+              child: const Text('Question'),
+            ),
+          ],
+        ),
+        InstabugButton(
+          onPressed: changeFloatingButtonEdge,
+          text: 'Move Floating Button to Left',
+        ),
+        InstabugButton(
+          onPressed: sendFeedback,
+          text: 'Send Feedback',
+        ),
+        InstabugButton(
+          onPressed: showNpsSurvey,
+          text: 'Show NPS Survey',
+        ),
+        InstabugButton(
+          onPressed: showManualSurvey,
+          text: 'Show Multiple Questions Survey',
+        ),
+        InstabugButton(
+          onPressed: showFeatureRequests,
+          text: 'Show Feature Requests',
+        ),
+        InstabugButton(
+          onPressed: _navigateToCrashes,
+          text: 'Crashes',
+        ),
+        InstabugButton(
+          onPressed: _navigateToApm,
+          text: 'APM',
+        ),
+        InstabugButton(
+          onPressed: _navigateToComplex,
+          text: 'Complex',
+        ),
+        const SectionTitle('Sessions Replay'),
+        InstabugButton(
+          onPressed: getCurrentSessionReplaylink,
+          text: 'Get current session replay link',
+        ),
+        const SectionTitle('Color Theme'),
+        ButtonBar(
+          mainAxisSize: MainAxisSize.max,
+          alignment: MainAxisAlignment.center,
+          children: <Widget>[
+            ElevatedButton(
+              onPressed: () => setColorTheme(ColorTheme.light),
+              style: ButtonStyle(
+                backgroundColor: MaterialStateProperty.all(Colors.white),
+                foregroundColor: MaterialStateProperty.all(Colors.lightBlue),
+              ),
+              child: const Text('Light'),
+            ),
+            ElevatedButton(
+              onPressed: () => setColorTheme(ColorTheme.dark),
+              style: ButtonStyle(
+                backgroundColor: MaterialStateProperty.all(Colors.black),
+                foregroundColor: MaterialStateProperty.all(Colors.white),
+              ),
+              child: const Text('Dark'),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+}
