+++ example/lib/src/widget/section_title.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class SectionTitle extends StatelessWidget {
+  final String text;
+
+  const SectionTitle(this.text, {Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      alignment: Alignment.centerLeft,
+      margin: const EdgeInsets.only(top: 20.0, left: 20.0),
+      child: Text(
+        text,
+        textAlign: TextAlign.left,
+        style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+      ),
+    );
+  }
+}
